FIX: clean app from unused files and functions
