+++ auth.ts
@@ -10,7 +10,6 @@ import { findUserByEmail, findUserById } from "./lib/utils";
 
 import type { NextAuthConfig } from "next-auth";
 import type { Role } from "@prisma/client";
-import { mergeAnnonymousBasketWithUserBasket } from "./lib/basket";
 
 export const authConfig = {
   providers: [
@@ -82,11 +81,6 @@ export const {
       token.role = currentUser.role;
 
       return token;
-    },
-  },
-  events: {
-    async signIn({ user }) {
-      await mergeAnnonymousBasketWithUserBasket(Number(user.id));
     },
   },
   adapter: PrismaAdapter(prisma),
