--- lib/hooks/useStore.ts
@@ -1,17 +1,0 @@
-import { useState, useEffect } from "react";
-
-const useStore = <T, F>(
-  store: (callback: (state: T) => unknown) => unknown,
-  callback: (state: T) => F,
-) => {
-  const result = store(callback) as F;
-  const [data, setData] = useState<F>();
-
-  useEffect(() => {
-    setData(result);
-  }, [result]);
-
-  return data;
-};
-
-export default useStore;
