+++ prisma/schema.prisma
@@ -24,7 +24,6 @@ model Product {
   approved      Boolean             @default(false)
   createdAt     DateTime            @default(now())
   updatedAt     DateTime            @updatedAt
-  BasketItem    BasketItem[]
 
   @@map("products")
 }
@@ -57,26 +56,6 @@ model CategoryToProduct {
   @@map("categoryToProduct")
 }
 
-model Basket {
-  id        Int          @id @default(autoincrement())
-  items     BasketItem[]
-  user      User?        @relation(fields: [userId], references: [id], onDelete: Cascade)
-  userId    Int?
-  createdAt DateTime     @default(now())
-  updatedAt DateTime     @updatedAt
-}
-
-model BasketItem {
-  id        Int      @id @default(autoincrement())
-  product   Product  @relation(fields: [productId], references: [id], onDelete: Cascade)
-  productId Int
-  Basket    Basket   @relation(fields: [basketId], references: [id], onDelete: Cascade)
-  basketId  Int
-  quantity  Int
-  createdAt DateTime @default(now())
-  updatedAt DateTime @updatedAt
-}
-
 enum Role {
   USER
   ADMIN
@@ -91,7 +70,6 @@ model User {
   password      String?
   role          Role      @default(USER)
   accounts      Account[]
-  Basket        Basket[]
 
   @@map("users")
 }
