+++ types/product.ts
@@ -28,9 +28,3 @@ export type ProductWithBrandAndCategoryNamePayload = Prisma.ProductGetPayload<{
     };
   };
 }>;
-
-export type BasketItemWithProduct = Prisma.BasketItemGetPayload<{
-  include: {
-    product: true;
-  };
-}>;
